pseiferth-libxtreemfsjava: Volume now uses Metadatacache. Implemented PeriodicThreads and FileInfo(partially).
git-svn-id: https://xtreemfs.googlecode.com/svn/branches/pseiferth-libxtreemfsjava@2750 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ java/servers/src/org/xtreemfs/common/libxtreemfs/Volume.java
@@ -313,8 +313,8 @@ public abstract class Volume {
      * 
      * @remark Ownership is transferred to the caller.
      */
-    public abstract DirectoryEntries readDir(UserCredentials userCredentials, String path, long offset,
-            long count, boolean namesOnly) throws IOException;
+    public abstract DirectoryEntries readDir(UserCredentials userCredentials, String path, int offset,
+            int count, boolean namesOnly) throws IOException;
 
     /**
      * Returns the list of extended attributes stored for "path" (Entries may be cached).
